requirements.txt
python-pptx==0.6.23
Flask==2.3.3
Werkzeug==2.3.7
